+++ src/adzerk/boot_reload/reload.cljs
@@ -5,7 +5,8 @@
    [goog.async.DeferredList :as deferred-list]
    [goog.net.jsloader       :as jsloader]))
 
-(def ^:private page-uri (goog.Uri. (.. js/window -location -href)))
+(defn- page-uri []
+  (goog.Uri. (.. js/window -location -href)))
 
 (defn- ends-with? [s pat]
   (= pat (subs s (- (count s) (count pat)))))
@@ -15,7 +16,7 @@
 
 (defn- normalize-href-or-uri [href-or-uri]
   (let [uri  (goog.Uri. href-or-uri)]
-    (.getPath (.resolve page-uri uri))))
+    (.getPath (.resolve (page-uri) uri))))
 
 (defn- changed-href? [href-or-uri changed]
   (when href-or-uri
@@ -41,10 +42,10 @@
                            :or {on-jsload identity}}]
   (let [js-files (filter #(ends-with? % ".js") changed)]
     (when (seq js-files)
-      (-> #(-> % guri/parse .makeUnique jsloader/load)
+      (-> #(-> % guri/parse .makeUnique)
         (map js-files)
         clj->js
-        deferred-list/gatherResults
+        (jsloader/loadMany)
         (.addCallbacks
           (fn [& _] (on-jsload))
           (fn [e] (.error js/console "Load failed:" (.-message e)))))
@@ -52,7 +53,7 @@
         (.trigger (js/jQuery js/document) "page-load")))))
 
 (defn- reload-html [changed]
-  (let [page-path (.getPath page-uri)
+  (let [page-path (.getPath (page-uri))
         html-path (if (ends-with? page-path "/")
                     (str page-path "index.html")
                     page-path)]
@@ -63,10 +64,19 @@
   (doseq [t things-to-log] (.log js/console t))
   (.groupEnd js/console))
 
+(defn has-dom?
+  "Perform heuristics to check if a non-shimmed DOM is available"
+  []
+  (and (exists? js/window)
+       (exists? js/window.document)
+       (exists? js/window.document.documentURI)))
+
 (defn reload [changed opts]
-  (group-log "Reload" changed)
-  (doto changed
-    (reload-js opts)
-    reload-html
-    reload-css
-    reload-img))
+  (let [changed* (map #(str (:asset-host opts) %) changed)]
+    (group-log "Reload" changed*)
+    (reload-js changed* opts)
+    (when (has-dom?)
+      (doto changed*
+        reload-html
+        reload-css
+        reload-img))))